+++ CHANGES.rst
@@ -18,6 +18,9 @@ New Features:
 - Solr 7 compatibility.
   [timo]
 
+- Index comments.
+  [gforcada]
+
 Bugfixes:
 
 - Make plone.protect dependency optional.